Build HTML structure
Add content for sidebar and main content. Sidebar includes a header,
navigation links, and temporary links for GitHub, LinkedIn, and email.
Main content includes an about section and a project section. The about
section consists of an intro, a summary, a list of skills, and a
temporary link to my resume. The project section consists of all
four projects and each of its respective description and temporary links
to the GitHub source and live GitHub site.