Build hamburger menu feature without burger icon
In index.html, add a new .sidebar-hamburger container that is initially
hidden but is shown when window narrows to below 'max-medium' breakpoint
of 800px. Meanwhile, hide .sidebar-top, .sidebar-middle, .sidebar-bottom
containers at the breakpoint.

Using JavaScript, toggle opening and closing of .sidebar-hamburger upon
clicking on the menu button.

The menu button is currently a pink button. The next step is to make the
hamburger menu icon that toggles between 3 bars and 'x' upon being
clicked.